+++ eng/scripts/Language-Settings.ps1
@@ -12,12 +12,23 @@ function Get-cpp-PackageInfoFromRepo($pkgPath, $serviceDirectory, $pkgName)
   # Test if the package path ends with the package name (e.g. sdk/storage/azure-storage-common)
   # This function runs in a loop where $pkgPath might be the path to the package and must return 
   # $null in cases where $pkgPath is not the path to the package specified by $pkgName
-  if ($pkgName -ne (Split-Path -Leaf $pkgPath)) { 
+  if ($pkgName -and ($pkgName -ne (Split-Path -Leaf $pkgPath))) { 
     return $null
   }
 
+  if (!$pkgName)
+  {
+    $pkgName = Split-Path -Leaf $pkgPath
+  }
+
   $packageVersion = & $PSScriptRoot/Get-PkgVersion.ps1 -ServiceDirectory $serviceDirectory -PackageName $pkgName
-  return [PackageProps]::new($pkgName, $packageVersion, $pkgPath, $serviceDirectory)
+  if ($null -ne $packageVersion)
+  {
+    $packageProps = [PackageProps]::new($pkgName, $packageVersion, $pkgPath, $serviceDirectory)
+    $packageProps.ArtifactName = $pkgName
+    return $packageProps
+  }
+  return $null
 }
 
 # Parse out package publishing information from a package-info.json file.