+++ sdk/storage/azure-storage-files-datalake/src/datalake_service_client.cpp
@@ -89,7 +89,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
                                       credential,
                                       Details::GetBlobClientOptions(options))
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(
         Storage::Details::ConstructPolicies(
@@ -110,7 +110,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
                                       credential,
                                       Details::GetBlobClientOptions(options))
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     Azure::Core::Http::TokenRequestOptions tokenOptions;
     tokenOptions.Scopes.emplace_back(Storage::Details::StorageScope);
@@ -132,7 +132,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
                                       Details::GetBlobUrlFromUrl(serviceUrl),
                                       Details::GetBlobClientOptions(options))
   {
-    Azure::Core::Http::TelemetryPolicyOptions telemetryPolicyOptions;
+    Azure::Core::Http::TelemetryOptions telemetryPolicyOptions;
     telemetryPolicyOptions.ApplicationId = options.ApplicationId;
     m_pipeline = std::make_shared<Azure::Core::Internal::Http::HttpPipeline>(
         Storage::Details::ConstructPolicies(